Do not swap text bytes in text_to_wkb_buf
text_to_wkb_buf was using bytes_to_wkb_buf, which swaps the byte order
according to the endianness. However, this is only needed for multi-byte
int or float types (and date/time types, which are large ints). Text
values are simply arrays of chars, so there should be no byte-swapping
going on. The initial int64 indicating the size of the string should
be swapped if needed, but this is correctly handled by int64_to_wkb_buf.

Also, we can avoid a call to text2cstring, which does an unnecessary
memory allocation. Use the VARSIZE_ANY_EXHDR and VARDATA macros to copy
from the text pointer directly.

Lastly, the old function was allocating space for the null terminator as
well, which is wrong, as we don't write the null terminator in the wkb
output. Fix this, and also correctly add the null terminator in
text_from_wkb_state. Palloc doesn't zero-out the memory, so this is
necessary or cstring2text could fail.