Fix the commutators of the ttext temporal compops (#722)
Fixes the issue mentioned by #680.

The commutators of the temporal compops for the `ttext` were wrong.
Previously this didn't raise an error, but starting with PG 17, postgres
raises an error when two operators have the same commutator. Strangely,
this error was only raised when running `pg_upgrade` to go from PG 16 to
PG 17, and does not get raised when running `CREATE INSTALL`.