Add upgrade file from 1.2.0 to 1.2.1 (#730)
The only SQL change between 1.2.0 and 1.2.1 are the commutator fixes
from 5e410dc86. This commit adds an upgrade file that can be used to go
from 1.2.0 to 1.2.1 by dropping and recreating the needed operators.
This file gets installed together with the rest of the files during
`make install`. To upgrade, run the following command after having build
and installed 1.2.1:

```
ALTER EXTENSION mobilitydb UPDATE TO "1.2.1";
```