Remove local and postgres includes from new exported header files (#748)
This PR just removes the local and postgres includes in two exportable
header files (makes PyMEOS build crash):

- `meos_pose.h`: removes `temporal/set.h`
- `meos_rgeo.h`: Removes `geo/stbox.h` and `postgres.h`.

For the postgres include in `meos_rgeo.h`, initially I tried to do it as
in `meos.h`, where the import is different depending on the `MEOS` flag
and then modified in the CMake, but it kept crashing when building
PyMEOS because the definitions where repeated. Since `meos_rgeo.h`
imports `meos.h`, I figured I could remove it altogether.